refactor math core
