+++ Cargo.toml
@@ -1,3 +1,6 @@
+[workspace]
+members = ["crates/*"]
+
 [package]
 name = "thyllore-animation"
 version = "0.0.1-preview"
@@ -16,6 +19,7 @@ path = "src/main.rs"
 crate-type = ["cdylib", "rlib"]
 
 [dependencies]
+thyllore-math-core = { path = "crates/thyllore-math-core" }
 anyhow = "1"
 log = "0.4"
 cgmath = "0.18"
