+++ crates/thyllore-math-core/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "thyllore-math-core"
+version = "0.0.1"
+edition = "2021"
+license = "Apache-2.0"
+authors = ["kodai731"]
+description = "Layer 0 pure math utilities for ThylloreAnimation (cgmath-based helpers, coordinate system, billboard, ray intersections)"
+repository = "https://github.com/kodai731/Thyllore-Animation"
+
+[dependencies]
+cgmath = "0.18"
+anyhow = "1"
