+++ src/math/mod.rs
@@ -1,22 +1,2 @@
-mod billboard;
-pub mod coordinate_system;
-mod matrix;
-mod quaternion;
-mod vector;
-
-pub use billboard::*;
-pub use coordinate_system::{
-    blender_to_world, fbx_to_world, fix_coord, get_camera_axes_from_view, gltf_to_world,
-    perspective, ray_plane_intersection, ray_to_line_segment_distance, ray_to_point_distance,
-    ray_to_triangle_barycentric, ray_to_triangle_intersection, screen_to_world_ray, view,
-    world_to_screen, world_y_axis, world_y_down,
-};
-pub use matrix::*;
-pub use quaternion::*;
-pub use vector::*;
-
-pub use cgmath::Quaternion;
-pub use cgmath::Rad;
-pub use cgmath::{point3, Deg, InnerSpace, MetricSpace, Vector2};
-pub use cgmath::{prelude::*, Vector3};
-pub use cgmath::{vec2, vec3, vec4};
+pub use thyllore_math_core::coordinate_system;
+pub use thyllore_math_core::*;
